feat(demo): restructure sample app
