+++ apps/demo/package.json
@@ -9,12 +9,11 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "vue": "3.5.32",
-    "vue-router": "5.0.4"
+    "vue": "3.5.32"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "6.0.6",
-    "sass": "^1.100.0",
+    "sass": "1.78.0",
     "vite": "5.4.21"
   }
 }
