+++ apps/demo/src/components/button.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts"></script>
+
+<template>
+  <button class="demo-button">测试按钮</button>
+</template>
+
+<style scoped lang="scss"></style>
