+++ apps/demo/src/styles/reset.scss
@@ -1,0 +1,5 @@
+* {
+  box-sizing: border-box;
+  padding: 0;
+  margin: 0;
+}
